requirements.txt
google-genai
python-dotenv
requests
mistralai